day 6

+++ src/pl/ksitarski/adventofcode/aoc2021/Day1Part1.java
@@ -10,7 +10,7 @@ public class Day1Part1 implements Solution {
     }
 
     @Override
-    public int solve(List<String> lines) {
+    public long solve(List<String> lines) {
         List<Integer> integers = lines.stream().map(Integer::parseInt).collect(Collectors.toList());
 
         int biggerThanPrevious = 0;
